src/inference/requirements.txt
kserve
minio
python-dotenv
